reorganize servers
